Move project handle form item to own dir
