+++ src/components/shared/formItems/index.tsx
@@ -4,7 +4,7 @@ import ProjectDescription from './ProjectDescription'
 import ProjectDiscord from './ProjectDiscord'
 import ProjectDiscountRate from './ProjectDiscountRate'
 import ProjectDuration from './ProjectDuration'
-import ProjectHandleFormItem from './ProjectHandle'
+import ProjectHandleFormItem from './ProjectHandle/ProjectHandleFormItem'
 import ProjectLink from './ProjectLink'
 import ProjectLogoUri from './ProjectLogoUri'
 import ProjectName from './ProjectName'
